+++ src/components/auth-required.test.ts
@@ -2,9 +2,9 @@ import { fireEvent } from '@testing-library/dom';
 import { html } from 'lit-html';
 import { mocked } from 'ts-jest/utils';
 import { fixture } from '../../__tests__/helpers/fixtures';
-import { SIGN_IN } from '../redux/constants';
 import { openDialog } from '../redux/dialogs/actions';
 import { store } from '../redux/store';
+import { SIGN_IN } from '../redux/user/types';
 import './auth-required';
 import { AuthRequired } from './auth-required';
