+++ src/redux/actions/index.ts
@@ -2,9 +2,8 @@ import firebase from 'firebase';
 import { helperActions } from './helper';
 import { sessionsActions } from './sessions';
 import { speakersActions } from './speakers';
-import { userActions } from './user';
 
-export { helperActions, sessionsActions, speakersActions, userActions };
+export { helperActions, sessionsActions, speakersActions };
 
 // TODO: Remove type
 declare global {
