+++ src/redux/constants.ts
@@ -18,7 +18,3 @@ export const FETCH_USER_FEATURED_SESSIONS_SUCCESS =
 export const SET_USER_FEATURED_SESSIONS = 'app/Sessions/SET_USER_FEATURED_SESSIONS';
 export const SET_USER_FEATURED_SESSIONS_FAILURE = 'app/Sessions/SET_USER_FEATURED_SESSIONS_FAILURE';
 export const SET_USER_FEATURED_SESSIONS_SUCCESS = 'app/Sessions/SET_USER_FEATURED_SESSIONS_SUCCESS';
-
-// user
-export const SIGN_IN = 'app/Auth/SIGN_IN';
-export const SIGN_OUT = 'app/Auth/SIGN_OUT';
