+++ src/redux/user/state.ts
@@ -1,0 +1,5 @@
+import { UserState } from './types';
+
+export const initialUserState: UserState = {
+  signedIn: false,
+};
